feat(Ban): 改进黑白名单页
