+++ package.json
@@ -21,13 +21,13 @@
     "clipboard": "^2.0.11",
     "codemirror": "^6.0.1",
     "cron-parser": "^4.9.0",
-    "dayjs": "^1.11.11",
+    "dayjs": "^1.11.12",
     "element-plus": "^2.7.7",
     "filesize": "^10.1.4",
     "lodash-es": "^4.17.21",
     "pinia": "^2.1.7",
     "randomcolor": "^0.6.2",
-    "vue": "^3.4.31",
+    "vue": "^3.4.33",
     "vue-diff": "^1.2.4",
     "vue-router": "^4.4.0",
     "vuedraggable": "^4.1.0"
@@ -36,21 +36,21 @@
     "@iconify-json/bi": "^1.1.23",
     "@iconify-json/carbon": "^1.1.36",
     "@types/lodash-es": "^4.17.12",
-    "@types/node": "^20.14.10",
+    "@types/node": "^20.14.11",
     "@types/randomcolor": "^0.5.9",
     "@typescript-eslint/eslint-plugin": "^6.21.0",
     "@typescript-eslint/parser": "^6.21.0",
     "@vitejs/plugin-legacy": "^5.4.1",
     "@vitejs/plugin-vue": "^5.0.5",
     "@vitejs/plugin-vue-jsx": "^4.0.0",
-    "@vue/compiler-sfc": "^3.4.31",
+    "@vue/compiler-sfc": "^3.4.33",
     "autoprefixer": "^10.4.19",
     "eslint": "^8.57.0",
     "eslint-plugin-vue": "^9.27.0",
     "iconify-icon": "^1.0.8",
     "postcss": "^8.4.39",
-    "postcss-nested": "^6.0.1",
-    "tailwindcss": "^3.4.5",
+    "postcss-nested": "^6.2.0",
+    "tailwindcss": "^3.4.6",
     "typescript": "^5.5.3",
     "unplugin-auto-import": "^0.17.8",
     "unplugin-icons": "^0.17.4",
